毎日の検索順位を取得するタスクをリファクタ
